feat(client): define type spotify api
