+++ client/src/pages/api/spotify/client/types/index.ts
@@ -1,0 +1,1 @@
+export * from "./result";
