+++ client/src/pages/api/spotify/client/types/result.ts
@@ -1,0 +1,3 @@
+export type OK<T> = { type: "ok"; value: T };
+export type Error<T> = { type: "error"; error: T };
+export type Result<T, U> = OK<T> | Error<U>;
